Move retryhandshake listener to listeners package
